backend/app/agents/coding/requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langgraph-checkpoint>=2.1.0
langgraph-prebuilt>=1.0.2
langgraph-sdk>=0.2.2
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-groq>=0.1.0
pydantic>=2.7.4,<3.0.0
requests>=2.31.0
anyio>=4.8.0
httpx>=0.28.1
python-dotenv>=1.0.0
zipfly>=0.1.0
pillow>=10.0.0
aiohttp>=3.9.0
typing-extensions>=4.8.0
openai>=1.109.1,<3.0.0
tiktoken>=0.7.0
xxhash>=3.5.0
ormsgpack>=1.12.0
# File parsing dependencies
PyPDF2>=3.0.0
python-docx>=0.8.11

# Note: pyarrow is optional (only needed for some Streamlit features)
# Install separately if needed: pip install pyarrow

